Break collect functions from services/__init__.py
These functions are unrelated to the service classes and are really
about managing the high-level control flow of issue aggregation.

As an additional benefit, this kills off what I believe are the last of
our ugly workarounds to avoid circular-dependencies (in db.py).
